python_youtube_player/requirements.txt
pafy==0.5.5
python-vlc==3.0.12118
youtube-dl==2021.4.26
